Break out flag types into their own package.
Make the GoDoc a bit easier to read.

This does make tests a bit harder, because any test is likely to want to
do some flag parsing, and we can't import the flag types into clif or
we'll get an import cycle. But that can be worked around by having a
basic string flag defined for the test code to use, which isn't a heavy
lift.