Move to a parser-based approach.
== Change how we route requests

Start implementing a parser-based approach, with multiple sweeps of the
input. This is a little easier to follow the logic, and the rules are
simplified.

First we do a pass just doing lexicographical parsing, marking what
things could be, or definitely are. For example, something could be a
command or an argument or a flag value in this pass, and we keep track
of those possibilities. But a flag key is usually pretty unmistakable as
a flag key, and so we keep track of knowing that for sure, too.

There's a normalization pass, as well, where we turn things like
`--flag=value` into a flag key (`--flag`) and a value (`value`) so the
input is consistent whether it was entered as `--flag=value` or
`--flag value`. We also find things that can only possibly be one thing
and assert that we know them to be that thing. Short flags like `-abc`
get turned into `-a -b -c`.

Finally, we bring the application structure into it and do a semantic
pass. Here, anything that matches an available subcommand of the current
command is considered a subcommand. This is simplified by the
realization that commands don't need both subcommands and arguments;
I've never seen a command that allowed both, and it leads to confusing
ambiguity. So if a command has subcommands, it cannot have arguments,
and we don't need to consider that the token may be an argument instead
of a subcommand. And if it's supposed to be a flag value, the user can
work around that pretty simply, either by using the `--flag=value`
syntax or moving it to literally any other point in the input. So that
seems pretty unambiguous.

Next, anything that could be a flag value is considered to be a flag
value. Yes, it could be an argument, but again, with the ambiguity being
what it is, a flag value is the easiest thing for a user to correct, by
moving the flag key or using an alternate syntax or just specifying a
flag value.

Finally, anything left that can be an argument is considered an
argument.

A normalization run is done again to clean up any lingering instances
where things are "unknown" but can only be one thing, and in theory
we've parsed the input.

== Update the flags strategy

Get rid of all the flag interfaces and cruft; we're probably headed
towards a more reflection-based interface to flags. Right now, we just
care about whether the flag had a value and, if so, what it was. We'll
handle the parsing of the value at the boundary between the framework
and the calling application.

Rename some options on FlagDef to have more reasonable defaults; the
empty value of a FlagDef should accept values, you shouldn't need to opt
out.

Add a bunch of validation, including configuration validation (to ensure
a valid Application has been defined) and validating that the flags the
user specified are real flags and meet the constraints of the flag
definitions on the Application.

As part of this, define a bunch of new error types and upgrade our
one-off errors into real errors.

Also, upgrade our flags and flag definitions. Give us the ability to
fall back on environment variables or hardcoded defaults for values.

Create a clif.FlagSet type we can use for automatic unmarshaling of
flags onto structs using clif.FlagSet.As(&foo). Making the
clif.FlagValues type that wraps []clif.FlagValue and using that on
clif.FlagSet also means we can normalize flag values into multiple flag
values, if that's something the end user desires. I still need to figure
out how I'm going to accept configuration for it, but having the ability
to automatically turn --flag=val1,val2,val3 into --flag=val1 --flag=val2
--flag=val3 in the eyes of the handling Application instead of
special-casing it each time seems nice. Especially when considered in
light of environment variables, where they can't be specified multiple
times.

Add an `As` method to the `clif.FlagValue`, `clif.FlagValues`, and
`clif.FlagSet` types that parses the flags into Go types.
`clif.FlagValue` can be parsed into string, bool, any int or uint type,
any float type, big.Int, big.Float, time.Time, or any type that
implements the `clif.FlagValueSetter` interface. `clif.FlagValues` can
be parsed into a slice, or if there's one or fewer values in the list,
that value can be parsed into anything `clif.FlagValue` supports, as a
special case. `clif.FlagSet` must be parsed into a map or a struct type.

🎵 Now Playing: Santa Tell Me
🎵 Artist: Sub-Radio
🎵 Album: Holiday Hangover
🎵 SpotifyID: 7MD5oB3r4pD4KyMiFH7z4e
🎵 ISRC: USL4Q1725881

🎵 Now Playing: Santa Claus Is Comin' to Town - Live at C.W. Post College, Greenvale, NY - December 1975
🎵 Artist: Bruce Springsteen
🎵 Album: Santa Claus Is Comin' to Town (Live at C.W. Post College, Greenvale, NY - December 1975)
🎵 SpotifyID: 6s2wpWPFPAgKg2LXxi1Oee
🎵 ISRC: USSM17500461

🎵 Now Playing: Rockin' Around The Christmas Tree
🎵 Artist: Brenda Lee
🎵 Album: Merry Christmas From Brenda Lee
🎵 SpotifyID: 2EjXfH91m7f8HiJN1yQg97
🎵 ISRC: USMC15848998

🎵 Now Playing: All I Want for Christmas Is You
🎵 Artist: Mariah Carey
🎵 Album: Merry Christmas
🎵 SpotifyID: 0bYg9bo50gSsH3LtXe2SQn
🎵 ISRC: USSM19400325

🎵 Now Playing: Santa Tell Me
🎵 Artist: Sub-Radio
🎵 Album: Holiday Hangover
🎵 SpotifyID: 7MD5oB3r4pD4KyMiFH7z4e
🎵 ISRC: USL4Q1725881

🎵 Now Playing: Have Yourself A Merry Little Christmas - Remastered 1999
🎵 Artist: Frank Sinatra
🎵 Album: A Jolly Christmas From Frank Sinatra
🎵 SpotifyID: 2FPfeYlrbSBR8PwCU0zaqq
🎵 ISRC: USCA29900549